refactor: extract callback module + add cache lock + docstring polish
Addresses reviewer feedback:
- Move callback helpers from background_review.py to background_review_callback.py
  for single-responsibility (background_review.py shouldn't also be a tool-call
  classifier + threading utility).
- Add threading.Lock around _callback_cache; cache failed loads to avoid re-importing.
- Clarify docstring: abstract message shape is test-only, runtime uses tool_calls + role:tool.
- Add comment on intentional patch/patch_file label normalization.
- Promote _WRITE_TOOL_NAMES constant.
- Strengthen test_callback_exception_does_not_propagate with explicit assertion.
